public/py_bills/requirements.txt
anyio==3.6.2
appdirs==1.4.4
argcomplete==1.10.3
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.8.2
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
camelot-py==0.9.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
compressed-rtf==1.0.6
contourpy==1.1.0
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
extract-msg==0.28.7
fake-useragent==1.1.3
fastapi==0.95.1
feedparser==6.0.10
fonttools==4.40.0
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
IMAPClient==2.1.0
importlib-metadata==6.8.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kaitaistruct==0.10
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
olefile==0.46
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.0
parse==1.19.1
parso==0.8.3
pdfminer.six==20191110
pdfminer3k==1.3.4
pdfplumber==0.9.0
pdfquery==0.4.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.2.0
playwright==1.32.1
ply==3.11
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pydivert==2.1.0
pyee==8.2.2
Pygments==2.15.0
pyhtml2pdf==0.0.6
PyMuPDF==1.22.3
PyMySQL==1.1.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==2.12.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2023.3
#pywin32==306
pyzmq==25.0.2
regex==2023.6.3
requests==2.28.2
requests-html==0.10.0
roman==4.0
selenium==4.8.3
selenium-wire==5.1.0
sgmllib3k==1.0.0
six==1.12.0
slate3k==0.5.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
stack-data==0.6.2
starlette==0.26.1
tabula-py==2.7.0
textblob==0.17.1
textract==1.6.5
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
TwoCaptcha==0.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
uvicorn==0.21.1
w3lib==2.1.1
Wand==0.6.11
wcwidth==0.2.6
webdriver-manager==3.8.6
webdrivermanager==0.10.0
websockets==10.4
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.0
yarg==0.1.9
zipp==3.16.0
zstandard==0.21.0
